@@ -5,52 +5,61 @@ import pipeactions = require("@aws-cdk/aws-codepipeline-actions");
 import s3 = require("@aws-cdk/aws-s3");
 import sns = require("@aws-cdk/aws-sns");
 import ecr = require("@aws-cdk/aws-ecr");
-import ecs = require("@aws-cdk/aws-ecs");
+import iam = require("@aws-cdk/aws-iam");
 
 export class ECSPipelineStack extends cdk.Stack {
   constructor(scope: cdk.App, id: string, props?: cdk.StackProps) {
     super(scope, id, props);
 
+    // Artifact bucket
     const artifactBucket = new s3.Bucket(this, "Artifact", {
       versioned: true
     });
 
     // Create an ECR resource and clean-up lifecycle rule
     const repository = new ecr.Repository(this, "ECR", {
-      repositoryName: "fargate-ecr"
+      repositoryName: "basicrepo"
     });
     repository.addLifecycleRule({ tagPrefixList: ["dev"], maxImageCount: 10 });
     repository.addLifecycleRule({ maxImageAgeDays: 7 });
 
+    // Code Build project - Build container and push to repo
     const buildProject = new codebuild.PipelineProject(this, "Build", {
       description: "Build project for the Fargate pipeline",
       environment: {
-        buildImage: codebuild.LinuxBuildImage.UBUNTU_14_04_DOCKER_18_09_0
+        buildImage: codebuild.LinuxBuildImage.UBUNTU_14_04_NODEJS_10_14_1,
+        privileged: true
       },
       buildSpec: {
         version: "0.2",
         phases: {
+          install: {
+            commands: ["echo Installing source NPM dependencies", "npm install"]
+          },
           pre_build: {
             commands: [
-              "echo logging in to AWS ECR...",
-              "$(aws ecr get -login --no-include-email --region ap-southeast-2)"
+              "echo Logging in to Amazon ECR...",
+              "$(aws ecr get-login --region $AWS_DEFAULT_REGION --no-include-email)",
+              "REPOSITORY_URI=375519225043.dkr.ecr.ap-southeast-2.amazonaws.com/basicrepo",
+              "COMMIT_HASH=$(echo $CODEBUILD_RESOLVED_SOURCE_VERSION | cut -c 1-7)",
+              "IMAGE_TAG=${COMMIT_HASH:=latest}"
             ]
           },
           build: {
             commands: [
-              "echo build Docker image on `date`",
-              "cd app",
-              "DOCKER_BUILDKIT=1 docker build --progress=plain -t go-app:latest .",
-              "docker tag go-app:latest $RepoName/go-app:latest"
+              "echo Build started on`date`",
+              "npm run build",
+              "echo Build the image",
+              "docker build -t $REPOSITORY_URI:latest .",
+              "docker tag $REPOSITORY_URI:latest $REPOSITORY_URI:$IMAGE_TAG"
             ]
           },
           post_build: {
-            commands: [
-              "echo build Docker image complete`date`",
-              "echo push latest Docker images to ECR...",
-              "docker push $RepoName/go-app:latest"
-            ]
+            commands: ["echo Build completed on`date`"]
           }
+        },
+        artifacts: {
+          files: ["**/*"]
         }
       }
     });
@@ -59,7 +68,7 @@ export class ECSPipelineStack extends cdk.Stack {
 
     const SourceAction = new pipeactions.S3SourceAction({
       actionName: "Source",
-      bucketKey: "/app/*",
+      bucketKey: "app.zip",
       bucket: artifactBucket,
       output: sourceOutput,
       pollForSourceChanges: true
@@ -78,7 +87,7 @@ export class ECSPipelineStack extends cdk.Stack {
         displayName: "ApprovalTopic",
         topicName: "PipelineApproval"
       }),
-      notifyEmails: ["<email>"]
+      notifyEmails: ["EMAIL"]
     });
 
     /*
@@ -95,6 +104,10 @@ export class ECSPipelineStack extends cdk.Stack {
       imageFile: buildOutput.atPath("imageDef.json")
     });
     */
+
+    const allowECR = new iam.PolicyStatement()
+      .addAction("ecr:*")
+      .addResource(repository.repositoryArn).allow;
 
     const pipeline = new codepipeline.Pipeline(this, "Pipeline", {
       pipelineName: "FargatePipeline",
@@ -113,5 +126,6 @@ export class ECSPipelineStack extends cdk.Stack {
         }
       ]
     });
+    //pipeline.addToRolePolicy(allowECR());
   }
 }
